extract boxing impl to its own module
